✨: 전역상태에 따른 Api 호출 커스텀 훅 구현 및 적용 완료, 정상 동작 확인
- 전역으로 관리되는 툴과 메니저 바 전역 상태에 따라 api호출을 관리하는 커스텀훅 구현 완료
